pyproject.toml
[project]
name = "fastapi-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "celery[redis]>=5.3.1",
    "fastapi[standard]>=0.135.1",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.13.1",
    "pyjwt[crypto]>=2.11.0",
    "redis[hiredis]>=7.3.0",
    "sqlalchemy>=2.0.48",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "anyio>=4.12.1",
    "httpx>=0.28.1",
    "locust>=2.43.3",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
]


[tool.pytest.ini_options]
markers = [
    "unit: unit tests",
    "integration: marks tests requiring external services",
    "functional: marks tests that need a database connection",
    "serial: marks tests that must run in series",
]
